task6

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -10,7 +10,7 @@
 int main(void)
 {
 
-	int num;
+long int num;
 
 num = 0;
